Get candidates totals

Aggregated candidate receipts and disbursements grouped by cycle.

Usage

fec_get_candidates_totals(
  office,
  min_receipts,
  cycle,
  district,
  max_receipts,
  sort_hide_null,
  state,
  max_debts_owed_by_committee,
  sort_nulls_last,
  min_debts_owed_by_committee,
  is_active_candidate,
  q,
  federal_funds_flag,
  per_page,
  max_cash_on_hand_end_period,
  max_disbursements,
  candidate_id,
  page,
  has_raised_funds,
  election_full,
  election_year,
  sort_null_only,
  min_cash_on_hand_end_period,
  sort,
  party,
  min_disbursements,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

office

Federal office candidate runs for: H, S or P

min_receipts

Minimum aggregated receipts

cycle

Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.

district

District of candidate

max_receipts

Maximum aggregated receipts

sort_hide_null

Hide null values on sorted column(s).

state

State of candidate

max_debts_owed_by_committee

Maximum debt

sort_nulls_last

Toggle that sorts null values last

min_debts_owed_by_committee

Minimum debt

is_active_candidate

Candidates who are actively seeking office. If no value is specified, all candidates are returned. When True is specified, only active candidates are returned. When False is specified, only inactive candidates are returned.

q

Name of candidate running for office

federal_funds_flag

A boolean the describes if a presidential candidate has accepted federal funds. The flag will be false for House and Senate candidates.

per_page

The number of results returned per page. Defaults to 20.

max_cash_on_hand_end_period

Maximum cash on hand

max_disbursements

Maximum aggregated disbursements

candidate_id

A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.

page

For paginating through results, starting at page 1

has_raised_funds

A boolean that describes if a candidate's committee has ever received any receipts for their campaign for this particular office. (Candidates have separate candidate IDs for each office.)

election_full

True indicates that full election period of a candidate. False indicates that two year election cycle.

election_year

Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.

sort_null_only

Toggle that filters out all rows having sort column that is non-null

min_cash_on_hand_end_period

Minimum cash on hand

sort

Provide a field to sort by. Use - for descending order.

party

Three-letter party code

min_disbursements

Minimum aggregated disbursements

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper