Get election dates

FEC election dates since 1995.

Usage

fec_get_election_dates(
  max_update_date,
  sort_nulls_last,
  sort_hide_null,
  max_primary_general_date,
  min_update_date,
  max_create_date,
  per_page,
  min_election_date,
  election_party,
  election_type_id,
  page,
  min_create_date,
  min_primary_general_date,
  election_year,
  election_district,
  office_sought,
  sort_null_only,
  max_election_date,
  sort,
  election_state,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

max_update_date

The maximum date this record was last updated.(MM/DD/YYYY or YYYY-MM-DD)

sort_nulls_last

Toggle that sorts null values last

sort_hide_null

Hide null values on sorted column(s).

max_primary_general_date

The maximum date of primary or general election.(MM/DD/YYYY or YYYY-MM-DD)

min_update_date

The minimum date this record was last updated.(MM/DD/YYYY or YYYY-MM-DD)

max_create_date

The maximum date this record was added to the system.(MM/DD/YYYY or YYYY-MM-DD)

per_page

The number of results returned per page. Defaults to 20.

min_election_date

The minimum date of election.

election_party

Party, if applicable.

election_type_id

Election type id

page

For paginating through results, starting at page 1

min_create_date

The minimum date this record was added to the system.(MM/DD/YYYY or YYYY-MM-DD)

min_primary_general_date

The minimum date of primary or general election.(MM/DD/YYYY or YYYY-MM-DD)

election_year

Year of election

election_district

House district of the office sought, if applicable.

office_sought

House, Senate or presidential office.

sort_null_only

Toggle that filters out all rows having sort column that is non-null

max_election_date

The maximum date of election.

sort

Provide a field to sort by. Use - for descending order.

election_state

State or territory of the office sought.

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper