Get legal search

Search legal documents by document type, or across all document types using keywords, parameter values and ranges.

Usage

fec_get_legal_search(
  hits_returned,
  af_report_year,
  case_max_open_date,
  ao_max_issue_date,
  case_statutory_citation,
  case_respondents,
  q,
  ao_min_issue_date,
  af_max_fd_date,
  from_hit,
  af_fd_fine_amount,
  type,
  af_name,
  ao_requestor_type,
  ao_statutory_citation,
  ao_entity_name,
  mur_type,
  ao_regulatory_citation,
  af_committee_id,
  ao_requestor,
  case_citation_require_all,
  af_min_fd_date,
  ao_is_pending,
  af_rtb_fine_amount,
  case_election_cycles,
  ao_category,
  ao_citation_require_all,
  case_dispositions,
  af_max_rtb_date,
  case_min_open_date,
  case_max_close_date,
  ao_min_request_date,
  ao_status,
  case_doc_category_id,
  af_min_rtb_date,
  ao_name,
  case_regulatory_citation,
  ao_no,
  case_min_close_date,
  sort,
  ao_max_request_date,
  case_no,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

hits_returned

Number of results to return (max 10)

af_report_year

Admin fine report year

case_max_open_date

The latest date opened of case

ao_max_issue_date

Latest issue date of advisory opinion

case_statutory_citation

Statutory citations

case_respondents

Cases respondents

q

Text to search legal documents for

ao_min_issue_date

Earliest issue date of advisory opinion

af_max_fd_date

The latest Final Determination date

from_hit

Get results starting from this index

af_fd_fine_amount

Final Determination fine amount

type

Legal Document type to refine search by - statutes - regulations - advisory_opinions - murs - admin_fines

af_name

Admin fine committee name

ao_requestor_type

Code of the advisory opinion requestor type.

ao_statutory_citation

Statutory citations

ao_entity_name

Name of commenter or representative

mur_type

Type of MUR : current or archived

ao_regulatory_citation

Regulatory citations

af_committee_id

Admin fine committee ID

ao_requestor

The requestor of the advisory opinion

case_citation_require_all

Require all citations to be in document (default behavior is any)

af_min_fd_date

The earliest Final Determination date

ao_is_pending

AO is pending

af_rtb_fine_amount

Reason to Believe fine amount

case_election_cycles

Cases election cycles

ao_category

Category of the document

ao_citation_require_all

Require all citations to be in document (default behavior is any)

case_dispositions

Cases dispositions

af_max_rtb_date

The latest Reason to Believe date

case_min_open_date

The earliest date opened of case

case_max_close_date

The latest date closed of case

ao_min_request_date

Earliest request date of advisory opinion

ao_status

Status of AO (pending, withdrawn, or final)

case_doc_category_id

Select one or more case_doc_category_id to filter by corresponding CASE_DOCUMENT_CATEGORY: - 1 - Conciliation Agreements - 2 - Complaint, Responses, Designation of Counsel and Extensions of Timee - 3 - General Counsel Reports, Briefs, Notifications and Responses - 4 - Certifications - 5 - Civil Penalties, Disgorgements and Other Payments - 6 - Statements of Reasons

af_min_rtb_date

The earliest Reason to Believe date

ao_name

Force advisory opinion name

case_regulatory_citation

Regulatory citations

ao_no

Force advisory opinion number

case_min_close_date

The earliest date closed of case

sort

Provide a field to sort by. Use - for descending order. ex: -case_no

ao_max_request_date

Latest request date of advisory opinion

case_no

Enforcement matter case number

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper