Get operations log

The Operations log contains details of each report loaded into the database. It is primarily used as status check to determine when all of the data processes, from initial entry through review are complete.

Usage

fec_get_operations_log(
  max_coverage_end_date,
  sort_nulls_last,
  amendment_indicator,
  sort_hide_null,
  max_receipt_date,
  per_page,
  min_transaction_data_complete_date,
  form_type,
  max_transaction_data_complete_date,
  beginning_image_number,
  page,
  min_coverage_end_date,
  sort,
  min_receipt_date,
  sort_null_only,
  candidate_committee_id,
  report_type,
  report_year,
  status_num,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

max_coverage_end_date

Ending date of the reporting period before this date(MM/DD/YYYY or YYYY-MM-DD)

sort_nulls_last

Toggle that sorts null values last

amendment_indicator

Amendent types: -N new -A amendment -T terminated -C consolidated -M multi-candidate -S secondary NULL might be new or amendment. If amendment indicator is null and the filings is the first or first in a chain treat it as if it was a new. If it is not the first or first in a chain then treat the filing as an amendment.

sort_hide_null

Hide null values on sorted column(s).

max_receipt_date

Selects all filings received before this date(MM/DD/YYYY or YYYY-MM-DD)

per_page

The number of results returned per page. Defaults to 20.

min_transaction_data_complete_date

Select all filings processed completely after this date(MM/DD/YYYY or YYYY-MM-DD)

form_type

The form where the underlying data comes from, for example, Form 1 would appear as F1: - F1 Statement of Organization - F1M Notification of Multicandidate Status - F2 Statement of Candidacy - F3 Report of Receipts and Disbursements for an Authorized Committee - F3P Report of Receipts and Disbursements by an Authorized Committee of a Candidate for The Office of President or Vice President - F3L Report of Contributions Bundled by Lobbyists/Registrants and Lobbyist/Registrant PACs - F3X Report of Receipts and Disbursements for other than an Authorized Committee - F4 Report of Receipts and Disbursements for a Committee or Organization Supporting a Nomination Convention - F5 Report of Independent Expenditures Made and Contributions Received - F6 48 Hour Notice of Contributions/Loans Received - F7 Report of Communication Costs by Corporations and Membership Organizations - F8 Debt Settlement Plan - F9 24 Hour Notice of Disbursements for Electioneering Communications - F13 Report of Donations Accepted for Inaugural Committee - F99 Miscellaneous Text - FRQ Request for Additional Information

max_transaction_data_complete_date

Select all filings processed completely before this date(MM/DD/YYYY or YYYY-MM-DD)

beginning_image_number

Unique identifier for the electronic or paper report. This number is used to construct PDF URLs to the original document.

page

For paginating through results, starting at page 1

min_coverage_end_date

Ending date of the reporting period after this date(MM/DD/YYYY or YYYY-MM-DD)

sort

Provide a field to sort by. Use - for descending order. ex: -case_no

min_receipt_date

Selects all filings received after this date(MM/DD/YYYY or YYYY-MM-DD)

sort_null_only

Toggle that filters out all rows having sort column that is non-null

candidate_committee_id

A unique identifier of the registered filer.

report_type

Name of report where the underlying data comes from: - 10D Pre-Election - 10G Pre-General - 10P Pre-Primary - 10R Pre-Run-Off - 10S Pre-Special - 12C Pre-Convention - 12G Pre-General - 12P Pre-Primary - 12R Pre-Run-Off - 12S Pre-Special - 30D Post-Election - 30G Post-General - 30P Post-Primary - 30R Post-Run-Off - 30S Post-Special - 60D Post-Convention - M1 January Monthly - M10 October Monthly - M11 November Monthly - M12 December Monthly - M2 February Monthly - M3 March Monthly - M4 April Monthly - M5 May Monthly - M6 June Monthly - M7 July Monthly - M8 August Monthly - M9 September Monthly - MY Mid-Year Report - Q1 April Quarterly - Q2 July Quarterly - Q3 October Quarterly - TER Termination Report - YE Year-End - ADJ COMP ADJUST AMEND - CA COMPREHENSIVE AMEND - 90S Post Inaugural Supplement - 90D Post Inaugural - 48 48 Hour Notification - 24 24 Hour Notification - M7S July Monthly/Semi-Annual - MSA Monthly Semi-Annual (MY) - MYS Monthly Year End/Semi-Annual - Q2S July Quarterly/Semi-Annual - QSA Quarterly Semi-Annual (MY) - QYS Quarterly Year End/Semi-Annual - QYE Quarterly Semi-Annual (YE) - QMS Quarterly Mid-Year/ Semi-Annual - MSY Monthly Semi-Annual (YE)

report_year

Forms with coverage date - year from the coverage ending date. Forms without coverage date - year from the receipt date.

status_num

Status of the transactional report. -0- Transaction is entered into the system. But not verified. -1- Transaction is verified.

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper