Get reports entity type

Each report represents the summary information from Form 3, Form 3X and Form 3P. These reports have key statistics that illuminate the financial status of a given committee. Things like cash on hand, debts owed by committee, total receipts, and total disbursements are especially helpful for understanding a committee's financial dealings. By default, this endpoint includes both amended and final versions of each report. To restrict to only the final versions of each report, use is_amended=false; to retrieve only reports that have been amended, use is_amended=true. Several different reporting structures exist, depending on the type of organization that submits financial information. To see an example of these reporting requirements, look at the summary and detailed summary pages of Form 3, Form 3X, and Form 3P. DISCLAIMER: The field labels contained within this resource are subject to change. We are attempting to succinctly label these fields while conveying clear meaning to ensure accessibility for all users.

Usage

fec_get_reports_entity_type(
  min_debts_owed_amount,
  max_debts_owed_expenditures,
  sort_hide_null,
  year,
  max_cash_on_hand_end_period_amount,
  filer_type,
  max_party_coordinated_expenditures,
  q_spender,
  max_receipt_date,
  per_page,
  max_independent_expenditures,
  min_party_coordinated_expenditures,
  committee_type,
  page,
  min_total_contributions,
  min_cash_on_hand_end_period_amount,
  min_receipt_date,
  sort_null_only,
  min_independent_expenditures,
  q_filer,
  max_disbursements_amount,
  max_total_contributions,
  cycle,
  amendment_indicator,
  sort_nulls_last,
  max_receipts_amount,
  is_amended,
  min_disbursements_amount,
  min_receipts_amount,
  candidate_id,
  beginning_image_number,
  sort,
  committee_id,
  report_type,
  most_recent,
  entity_type,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

min_debts_owed_amount

Filter for all amounts greater than a value.

max_debts_owed_expenditures

Filter for all amounts less than a value.

sort_hide_null

Hide null values on sorted column(s).

year

Forms with coverage date - year from the coverage ending date. Forms without coverage date - year from the receipt date.

max_cash_on_hand_end_period_amount

Filter for all amounts less than a value.

filer_type

The method used to file with the FEC, either electronic or on paper.

max_party_coordinated_expenditures

Filter for all amounts less than a value.

q_spender

Keyword search for spender name or ID

max_receipt_date

Selects all items received by FEC before this date(MM/DD/YYYY or YYYY-MM-DD)

per_page

The number of results returned per page. Defaults to 20.

max_independent_expenditures

Filter for all amounts less than a value.

min_party_coordinated_expenditures

Filter for all amounts greater than a value.

committee_type

The one-letter type code of the organization: - C communication cost - D delegate - E electioneering communication - H House - I independent expenditure filer (not a committee) - N PAC - nonqualified - O independent expenditure-only (super PACs) - P presidential - Q PAC - qualified - S Senate - U single candidate independent expenditure - V PAC with non-contribution account, nonqualified - W PAC with non-contribution account, qualified - X party, nonqualified - Y party, qualified - Z national party non-federal account

page

For paginating through results, starting at page 1

min_total_contributions

Filter for all amounts greater than a value.

min_cash_on_hand_end_period_amount

Filter for all amounts greater than a value.

min_receipt_date

Selects all items received by FEC after this date(MM/DD/YYYY or YYYY-MM-DD)

sort_null_only

Toggle that filters out all rows having sort column that is non-null

min_independent_expenditures

Filter for all amounts greater than a value.

q_filer

Keyword search for filer name or ID

max_disbursements_amount

Filter for all amounts less than a value.

max_total_contributions

Filter for all amounts less than a value.

cycle

Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.

amendment_indicator

Amendent types: -N new -A amendment -T terminated -C consolidated -M multi-candidate -S secondary NULL might be new or amendment. If amendment indicator is null and the filings is the first or first in a chain treat it as if it was a new. If it is not the first or first in a chain then treat the filing as an amendment.

sort_nulls_last

Toggle that sorts null values last

max_receipts_amount

Filter for all amounts less than a value.

is_amended

False indicates that a report is the most recent. True indicates that the report has been superseded by an amendment.

min_disbursements_amount

Filter for all amounts greater than a value.

min_receipts_amount

Filter for all amounts greater than a value.

candidate_id

A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.

beginning_image_number

Unique identifier for the electronic or paper report. This number is used to construct PDF URLs to the original document.

sort

Provide a field to sort by. Use - for descending order. ex: -case_no

committee_id

A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.

report_type

Report type; prefix with "-" to exclude. Name of report where the underlying data comes from: - 10D Pre-Election - 10G Pre-General - 10P Pre-Primary - 10R Pre-Run-Off - 10S Pre-Special - 12C Pre-Convention - 12G Pre-General - 12P Pre-Primary - 12R Pre-Run-Off - 12S Pre-Special - 30D Post-Election - 30G Post-General - 30P Post-Primary - 30R Post-Run-Off - 30S Post-Special - 60D Post-Convention - M1 January Monthly - M10 October Monthly - M11 November Monthly - M12 December Monthly - M2 February Monthly - M3 March Monthly - M4 April Monthly - M5 May Monthly - M6 June Monthly - M7 July Monthly - M8 August Monthly - M9 September Monthly - MY Mid-Year Report - Q1 April Quarterly - Q2 July Quarterly - Q3 October Quarterly - TER Termination Report - YE Year-End - ADJ COMP ADJUST AMEND - CA COMPREHENSIVE AMEND

most_recent

Report is either new or is the most-recently filed amendment

entity_type

Committee groupings based on FEC filing form. Choose one of: presidential, pac-party, house-senate, or ie-only

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper