Get schedules schedule a by state by candidate totals

Itemized individual contributions aggregated by contributor’s state, candidate, committee type and cycle. If you are interested in our “is_individual” methodology, review the methodology page. Unitemized individual contributions are not included.

Usage

fec_get_schedules_schedule_a_by_state_by_candidate_totals(
  candidate_id,
  sort_nulls_last,
  cycle,
  page,
  sort_hide_null,
  election_full,
  sort_null_only,
  per_page,
  sort,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

candidate_id

A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.

sort_nulls_last

Toggle that sorts null values last

cycle

Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.

page

For paginating through results, starting at page 1

sort_hide_null

Hide null values on sorted column(s).

election_full

True indicates that full election period of a candidate. False indicates that two year election cycle.

sort_null_only

Toggle that filters out all rows having sort column that is non-null

per_page

The number of results returned per page. Defaults to 20.

sort

Provide a field to sort by. Use - for descending order.

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper