Get schedules schedule e

Schedule E covers the line item expenditures for independent expenditures. For example, if a super PAC bought ads on TV to oppose a federal candidate, each ad purchase would be recorded here with the expenditure amount, name and id of the candidate, and whether the ad supported or opposed the candidate. An independent expenditure is an expenditure for a communication "expressly advocating the election or defeat of a clearly identified candidate that is not made in cooperation, consultation, or concert with, or at the request or suggestion of, a candidate, a candidate’s authorized committee, or their agents, or a political party or its agents." Aggregates by candidate do not include 24 and 48 hour reports. This ensures we don't double count expenditures and the totals are more accurate. You can still find the information from 24 and 48 hour reports in /schedule/schedule_e/. Due to the large quantity of Schedule E filings, this endpoint is not paginated by page number. Instead, you can request the next page of results by adding the values in the last_indexes object from pagination to the URL of your last request. For example, when sorting by expenditure_amount, you might receive a page of results with the following pagination information: &quot;pagination&quot;: { &quot;count&quot;: 152623, &quot;last_indexes&quot;: { &quot;last_index&quot;: &quot;3023037&quot;, &quot;last_expenditure_amount&quot;: -17348.5 }, &quot;per_page&quot;: 20, &quot;pages&quot;: 7632 } } To fetch the next page of sorted results, append last_index=3023037 and last_expenditure_amount= to the URL. We strongly advise paging through these results by using the sort indices (defaults to sort by disbursement date, e.g. last_disbursement_date), otherwise some resources may be unintentionally filtered out. This resource uses keyset pagination to improve query performance and these indices are required to properly page through this large dataset. Note: because the Schedule E data includes many records, counts for large result sets are approximate; you will want to page through the records until no records are returned.

Usage

fec_get_schedules_schedule_e(
  max_dissemination_date,
  sort_hide_null,
  payee_name,
  q_spender,
  per_page,
  last_index,
  min_dissemination_date,
  candidate_office_state,
  filing_form,
  sort_null_only,
  max_amount,
  max_image_number,
  min_filing_date,
  cycle,
  sort_nulls_last,
  support_oppose_indicator,
  candidate_office,
  is_notice,
  line_number,
  last_expenditure_amount,
  last_expenditure_date,
  candidate_id,
  last_office_total_ytd,
  last_support_oppose_indicator,
  min_date,
  candidate_party,
  committee_id,
  min_image_number,
  max_filing_date,
  candidate_office_district,
  min_amount,
  image_number,
  sort,
  max_date,
  most_recent,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

max_dissemination_date

Selects all items distributed by this committee before this date

sort_hide_null

Hide null values on sorted column(s).

payee_name

Name of the entity that received the payment.

q_spender

Keyword search for spender name or ID

per_page

The number of results returned per page. Defaults to 20.

last_index

Index of last result from previous page

min_dissemination_date

Selects all items distributed by this committee after this date

candidate_office_state

US state or territory

filing_form

The form where the underlying data comes from, for example, Form 1 would appear as F1: - F1 Statement of Organization - F1M Notification of Multicandidate Status - F2 Statement of Candidacy - F3 Report of Receipts and Disbursements for an Authorized Committee - F3P Report of Receipts and Disbursements by an Authorized Committee of a Candidate for The Office of President or Vice President - F3L Report of Contributions Bundled by Lobbyists/Registrants and Lobbyist/Registrant PACs - F3X Report of Receipts and Disbursements for other than an Authorized Committee - F4 Report of Receipts and Disbursements for a Committee or Organization Supporting a Nomination Convention - F5 Report of Independent Expenditures Made and Contributions Received - F6 48 Hour Notice of Contributions/Loans Received - F7 Report of Communication Costs by Corporations and Membership Organizations - F8 Debt Settlement Plan - F9 24 Hour Notice of Disbursements for Electioneering Communications - F13 Report of Donations Accepted for Inaugural Committee - F99 Miscellaneous Text - FRQ Request for Additional Information

sort_null_only

Toggle that filters out all rows having sort column that is non-null

max_amount

Filter for all amounts less than a value.

max_image_number

Maxium image number of the page where the schedule item is reported

min_filing_date

Selects all filings received after this date

cycle

Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.

sort_nulls_last

Toggle that sorts null values last

support_oppose_indicator

Explains if the money was spent in order to support or oppose a candidate or candidates. (Coded S or O for support or oppose.) This indicator applies to independent expenditures and communication costs.

candidate_office

Federal office candidate runs for: H, S or P

is_notice

Record filed as 24- or 48-hour notice.

line_number

Filter for form and line number using the following format: FORM-LINENUMBER. For example an argument such as F3X-16 would filter down to all entries from form F3X line number 16.

last_expenditure_amount

When sorting by expenditure_amount, this is populated with the expenditure_amount of the last result. However, you will need to pass the index of that last result to last_index to get the next page.

last_expenditure_date

When sorting by expenditure_date, this is populated with the expenditure_date of the last result. However, you will need to pass the index of that last result to last_index to get the next page.

candidate_id

A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.

last_office_total_ytd

When sorting by office_total_ytd, this is populated with the office_total_ytd of the last result. However, you will need to pass the index of that last result to last_index to get the next page.'

last_support_oppose_indicator

When sorting by support_oppose_indicator, this is populated with the support_oppose_indicator of the last result. However, you will need to pass the index of that last result to last_index to get the next page.'

min_date

Minimum date

candidate_party

Three-letter code for the party affiliated with a candidate or committee. For example, DEM for Democratic Party and REP for Republican Party.

committee_id

A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.

min_image_number

Minium image number of the page where the schedule item is reported

max_filing_date

Selects all filings received before this date

candidate_office_district

Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.

min_amount

Filter for all amounts greater than a value.

image_number

An unique identifier for each page where the electronic or paper filing is reported.

sort

Provide a field to sort by. Use - for descending order.

max_date

Maximum date

most_recent

The report associated with the transaction is either new or is the most-recently filed amendment. Undetermined version (null) is always included.

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper