Get schedules schedule e efile

Efiling endpoints provide real-time campaign finance data received from electronic filers. Efiling endpoints only contain the most recent four months of data and don't contain the processed and coded data that you can find on other endpoints.

Usage

fec_get_schedules_schedule_e_efile(
  spender_name,
  min_expenditure_date,
  sort_nulls_last,
  max_dissemination_date,
  support_oppose_indicator,
  sort_hide_null,
  candidate_office,
  is_notice,
  payee_name,
  per_page,
  candidate_search,
  max_expenditure_amount,
  min_dissemination_date,
  min_filed_date,
  candidate_id,
  candidate_office_state,
  max_filed_date,
  page,
  min_expenditure_amount,
  committee_id,
  candidate_party,
  max_expenditure_date,
  filing_form,
  sort_null_only,
  candidate_office_district,
  image_number,
  sort,
  most_recent,
  api_key = Sys.getenv("FEC_API_KEY", unset = "DEMO_KEY")
)

Arguments

spender_name

The name of the committee. If a committee changes its name, the most recent name will be shown. Committee names are not unique. Use committee_id for looking up records.

min_expenditure_date

Selects all items expended by this committee after this date

sort_nulls_last

Toggle that sorts null values last

max_dissemination_date

Selects all items distributed by this committee before this date

support_oppose_indicator

Explains if the money was spent in order to support or oppose a candidate or candidates. (Coded S or O for support or oppose.) This indicator applies to independent expenditures and communication costs.

sort_hide_null

Hide null values on sorted column(s).

candidate_office

Federal office candidate runs for: H, S or P

is_notice

Record filed as 24- or 48-hour notice.

payee_name

Name of the entity that received the payment.

per_page

The number of results returned per page. Defaults to 20.

candidate_search

Search for candidates by candiate id or candidate first or last name

max_expenditure_amount

Selects all items expended by this committee less than this amount

min_dissemination_date

Selects all items distributed by this committee after this date

min_filed_date

Timestamp of electronic or paper record that FEC received

candidate_id

A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.

candidate_office_state

US state or territory where a candidate runs for office

max_filed_date

Timestamp of electronic or paper record that FEC received

page

For paginating through results, starting at page 1

min_expenditure_amount

Selects all items expended by this committee greater than this amount

committee_id

A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.

candidate_party

Three-letter code for the party affiliated with a candidate or committee. For example, DEM for Democratic Party and REP for Republican Party.

max_expenditure_date

Selects all items expended by this committee before this date

filing_form

The form where the underlying data comes from, for example, Form 1 would appear as F1: - F1 Statement of Organization - F1M Notification of Multicandidate Status - F2 Statement of Candidacy - F3 Report of Receipts and Disbursements for an Authorized Committee - F3P Report of Receipts and Disbursements by an Authorized Committee of a Candidate for The Office of President or Vice President - F3L Report of Contributions Bundled by Lobbyists/Registrants and Lobbyist/Registrant PACs - F3X Report of Receipts and Disbursements for other than an Authorized Committee - F4 Report of Receipts and Disbursements for a Committee or Organization Supporting a Nomination Convention - F5 Report of Independent Expenditures Made and Contributions Received - F6 48 Hour Notice of Contributions/Loans Received - F7 Report of Communication Costs by Corporations and Membership Organizations - F8 Debt Settlement Plan - F9 24 Hour Notice of Disbursements for Electioneering Communications - F13 Report of Donations Accepted for Inaugural Committee - F99 Miscellaneous Text - FRQ Request for Additional Information

sort_null_only

Toggle that filters out all rows having sort column that is non-null

candidate_office_district

Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.

image_number

An unique identifier for each page where the electronic or paper filing is reported.

sort

Provide a field to sort by. Use - for descending order.

most_recent

The report associated with the transaction is either new or is the most-recently filed amendment. Undetermined version (null) is always included.

api_key

An API key provided by the API provider. This key is not clearly documented in the API description. Check the API documentation for details.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper